Create radialItems.ts utility file to manage radial data assignment
